battery/AC check
I'm thinking mprg could be configured to disable post-processing when on
battery power. to preserve battery life, sure... but I'm also thinking it might
often be the case that I'm on AC-power when "composing", and on battery-power
when recording. works both ways methinks: post-processing is for staring deep
into the screen and being inspired; whereas recording is more of an engineering
thing, where I just need to "press space to record", and I definitely don't
want my laptop fans adding noise to a recording :)